requirements_1.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.0.5
django-mongoengine==0.4.4
django-rest-framework-mongoengine==3.4.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.4
djoser==2.1.0
dnspython==2.1.0
ffmpeg==1.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
mutagen==1.45.1
numpy==1.20.0
oauthlib==3.1.0
opencv-python==4.5.1.48
Pillow==8.1.0
pkg-resources==0.0.0
pycparser==2.20
pydub==0.24.1
PyJWT==2.0.1
pymongo==3.11.3
pytesseract==0.3.7
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
SpeechRecognition==3.8.1
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.3
